use mysqli
This is not a stable release (or a point release). It is a work in progress. Please test it and report any issues.

a proper readme will follow soon

we are at an odd stage where we need both mysqli and pdo

We will move to pdo but this way we can do this in a piecemeal fashion